Week 4 first draft
This is as far as I’ve got. I have a suspicion now that the photo
should be on the background of the whole page but I probably won’t have
time to try that. I hope to still have time to add the bottom section
and footer before class.